feat(ci): Update GitHub workflows to use larger Ubuntu runner (#242)
* Update GitHub workflows to use larger Ubuntu runner

In this commit, the OS specified in both the pipeline.yml and docker.yml GitHub workflows has been changed from "ubuntu-latest" to "ubuntu-latest-large". This will enable larger compute resources for our workflow jobs.

* Update runs-on syntax in GitHub workflows

The old syntax for specifying the runner within GitHub workflows has been replaced with a new one. Instead of directly setting `runs-on`, we are required to specify `runs-on` with `labels`. Applied to both `pipeline.yml` and `docker.yml` workflows.

---------

Co-authored-by: Devon Bear <itsdevbear@berachain.com>